% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream-in.r
\name{warc_stream_in}
\alias{warc_stream_in}
\title{Stream in records from a WARC file}
\usage{
warc_stream_in(path, handler, ..., warc_types = NULL)
}
\arguments{
\item{path}{path to WARC file}

\item{handler}{callback function to call for each record}

\item{...}{optional arguments to \code{handler}}

\item{warc_types}{if provided, only WARC record types matching the ones specified will
be streamed in and passed to \code{f()}. Valid options are: \code{warcinfo}, \code{request},
\code{response}, \code{resource}, \code{metadata}, \code{revisit}, \code{conversion} then returned WARC
records.}
}
\value{
the number of records processed (invisibly)
}
\description{
This is a pure R function that streams in WARC records and calls a callback function
with the WARC headers and payload for each record, optionally filtering by a subset
of WARC record types.
}
\details{
The signature of the callback function should be:

function(headers, payload, ...) { }
}
\examples{
myfun <- function(headers, payload, ...) {
  print(as.numeric(headers$`content-length`) == length(payload))
}

warc_stream_in(
  system.file("extdata/sample.warc.gz", package="jwatr"),
  myfun
)
}
