% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-response.r
\name{warc_write_response}
\alias{warc_write_response}
\title{Write simple \code{httr::GET} requests or full \code{httr} \code{response} objects to a WARC file}
\usage{
warc_write_response(wobj, x)
}
\arguments{
\item{wobj}{a WARC file object created with \url{warc_file}}

\item{x}{Either a single, valid URL to retrieve the contents of or an \code{httr} \code{response} object}
}
\value{
The passed-in \code{wobj} (thus, allowing for pipe chaining)
}
\description{
If a plain, single, character URL is provided, a \code{curl} request will be made
and the contents of the result will be added to the WARC file.
}
\details{
If an \code{httr} \code{response} object is provided, it will be transformed and the
contents of the result will be added to the WARC file.
}
\examples{
\dontrun{
tf <- tempfile("test")
wf <- warc_file(tf)
warc_write_response(wf, "https://rud.is/b/")
warc_write_response(wf, GET("https://rud.is/b/"))
warc_write_response(wf, "https://www.rstudio.com/")
warc_write_response(wf, "https://www.r-project.org/")
warc_write_response(wf, "http://che.org.il/wp-content/uploads/2016/12/pdf-sample.pdf")

POST(
  url = "https://data.police.uk/api/crimes-street/all-crime",
  query = list( lat = "52.629729", lng = "-1.131592", date = "2017-01")
) -> uk_res

warc_write_response(wf, uk_res)
warc_write_response(wf, "https://journal.r-project.org/RLogo.png")

close_warc_file(wf)
unlink(tf)
}
}
