% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-warcinfo.r
\name{warc_write_warcinfo}
\alias{warc_write_warcinfo}
\title{Write a \code{warcinfo} record to a WARC File}
\usage{
warc_write_warcinfo(wobj, warc_date = Sys.time(), warc_record_id = NULL,
  warc_info = list(software = sprintf("jwatr \%s", packageVersion("jwatr")),
  format = "WARC File Format 1.0"))
}
\arguments{
\item{wobj}{a WARC file object created with \url{warc_file}}

\item{warc_date}{A supplied \code{POSIXct} timestamp to use to timestamp the WARC file.
Current time will be used if none supplied.}

\item{warc_record_id}{A unique identifier for the WARC record. If not provided one
will be generated with \code{UUIDgenerate}.}

\item{warc_info}{A named \code{list} with fields to store for the \code{warcinfo} record.}
}
\value{
The passed-in \code{wobj} (thus, allowing for pipe chaining)
}
\description{
Write a \code{warcinfo} record to a WARC File
}
