% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jwatr-package.R
\docType{package}
\name{jwatr}
\alias{jwatr}
\alias{jwatr-package}
\title{Tools to Query and Create Web Archive Files Using the Java Web Archive Toolkit}
\description{
The Java Web Archive Toolkit ('JWAT') \url{https://sbforge.org/display/JWAT/Overview}
is a library of Java objects and methods which enables reading, writing and validating web
archive files.
}
\author{
Bob Rudis (bob@rud.is)
}
