% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warc-file.r
\name{warc_write_response}
\alias{warc_write_response}
\title{Add the response to a HTTP GET request to a WARC file}
\usage{
warc_write_response(wobj, URL)
}
\arguments{
\item{wobj}{a WARC file object created with [warc_file]()}

\item{URL}{A valid URL to retrieve the contents of}
}
\description{
Add the response to a HTTP GET request to a WARC file
}
\examples{
\dontrun{
wf <- warc_file("~/Desktop/test")
warc_write_response(wf, "https://rud.is/b/")
warc_write_response(wf, "https://www.rstudio.com/")
warc_write_response(wf, "https://www.r-project.org/")
warc_write_response(wf, "https://journal.r-project.org/archive/2016-2/RJ-2016-2.pdf")
warc_write_response(wf, "https://journal.r-project.org/RLogo.png")
close_warc_file(wf)
}
}
